requirements.txt
python-frontmatter
newsapi
python-twitter
green
PyCryptoDome
pylint
sqlalchemy
green
PyCryptoDome
pylint
sqlalchemy
chardet>=2.3.0
entropy>=0.9
klein>=16.12.0
patch>=1.16
profanity>=1.1
py2neo>=3.1.2
pycld2>=0.31
pycryptodome>=3.4.3
py-gfm>=0.1.3
pyjwt>=1.4.2
python-vote-full>=1.0
pyyaml>=3.11
redlock>=1.2.0
stem>=1.5.3
twisted>=16.6.0
txredisapi>=1.4.4
unidiff
urwid>=1.3.1
# python 3 support needed
#progressbar>=2.3
progressbar33>=2.3
Flask==0.12.2
Flask-Login==0.4.0
Flask-Session==0.3.1
Flask_SocketIO
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
python-engineio
python-socketio
six==1.9.0
Werkzeug==0.12.2
